+++ app/src/main/java/io/numberrun/Component/Renderable.java
@@ -19,11 +19,11 @@ public interface Renderable extends Component {
      *
      * @return 描画順序
      */
-    default int getZOrder() {
+    default float getZOrder() {
         return 0;
     }
 
-    void setZOrder(int zOrder);
+    void setZOrder(float zOrder);
 
     default float getWidth() {
         return 0;
@@ -33,7 +33,7 @@ public interface Renderable extends Component {
         return 0;
     }
 
-    default Renderable withZOrder(int zOrder) {
+    default Renderable withZOrder(float zOrder) {
         setZOrder(zOrder);
         return this;
     }
